rag-groq-chroma/requirements.txt
# Environment
python-dotenv

# HuggingFace
datasets
sentence-transformers
torch

# Vector DB
chromadb

# LLM
groq

# Utils
tqdm
